Time: 1 ms (95.85%), Space: 50.3 MB (33.29%) - LeetHub

+++ 136-single-number/136-single-number.java
@@ -1,7 +1,7 @@
 class Solution {
     public int singleNumber(int[] nums) {
         short number = (short)nums[0];
-        for(int i=1;i<nums.length;i++) number ^= (short)nums[i];
+        for(int i=1;i<nums.length;i++) number ^= nums[i];
         return number;
     }
 }
